Enable nullable on demos, and deal with consequences
Enabling nullable on the pet store projects revealed that we weren't yet done. In particular, when passing parameters around for resolving links, it turned out that we needed the be able to pass (name, null) pairs in. For parameters that turn into query string arguments (e.g., continuation values in the pet store demo) we explicitly pass in an argument indicating that we don't want that one. (This isn't technically necessary—you can also just omit it. But it's convenient to be able to pass null.)