Drop netstandard2.0 support
Since we're bumping the major version number, this is a good time to do this. .NET Core 3.1 has been around for a long time now, and we've not built anything new for .NET Core 2.1 for a long time. (And if necessary, we can always release 2.x updates.)

This enables us to avoid the hacky direct dependencies on some V2.2 ASP.NET Core components. We can now use the official supported approach of using a FrameworkReference to get access to ASP.NET Core components.