Improve testing for converters (#316)
This revealed that the ConvertTo for these methods didn't seem to work at all. (It would invariably fail when trying to validate its converted result, because it creates a string, but validation for these types fails unless the input is of the corresponding JToken type.)

Also, there were inconsistencies in how validation failures were being reported, which this fixes.

Note that this now uses the .NET 7 SDK in build pipeline (but only for access to new language features - there's no change in target runtimes).